add device rotation cases to user journey acceptance tests
